packing list error
